+++ packages/orq-rc/RELEASES.md
@@ -10369,3 +10369,13 @@ Based on:
 - [python v4.14.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.42 - packages/orq-rc
+
+## 2026-08-20 18:45:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.0 (2.931.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.43 - packages/orq-rc